5. Конвертер валюты [#41583 #188090] изменил тип переменных

+++ src/ru/job4j/converter/Converter.java
@@ -2,12 +2,12 @@ package ru.job4j.converter;
 
 public class Converter {
 
-    public static double rubleToEuro(int value) {
-        return (double) value / 70;
+    public static float rubleToEuro(float value) {
+        return value / 70;
     }
 
-    public static double rubleToDollar(int value) {
-        return (double) value / 60;
+    public static float rubleToDollar(float value) {
+        return  value / 60;
     }
 
     public static void main(String[] args) {
